requirements.txt
pandas
mysql-connector-python
apache-airflow